requirements.txt
streamlit==1.39.0
requests==2.32.3
folium==0.18.0
streamlit-folium==0.23.1
geopy==2.4.1
pandas==2.2.3
numpy==2.1.2
plotly==5.24.1
python-dateutil==2.8.2
certifi==2024.2.2
qrcode==7.4.2
Pillow==10.2.0
opencv-python-headless==4.9.0.80
streamlit-qrcode-scanner
reportlab==4.0.8
pillow>=9.0.0
streamlit-drawable-canvas==0.9.3
